Mobile App Design
Mobile App Design Complete.
Design Consist in 2 parts, the first part is the login for the App, now
works just with a hardcode password, but will be implemented with
database.
Second part is a list of products, now data is extracted by hardcode
array items in typescript file, but will be extracted from database.